Add delete and clear all completed functions
