+++ src/index.html
@@ -13,7 +13,7 @@
             <button id="submit-button"><i class="fas fa-check"></i></button>
         </form>
         <ul id="to-do-list"></ul>
-        <button class="to-do-clear">Clear all completed</button>
+        <button id="to-do-clear">Clear all completed</button>
     </section>
 </body>
 </html>
