+++ src/style.css
@@ -50,14 +50,27 @@ i { pointer-events: none; }
 
 .svg-inline--fa { color: rgb(185, 185, 185); }
 
-#submit-button,
-.move-button {
+button {
   border: 0;
   background-color: inherit;
   margin-right: 0.5em;
 }
 
 #submit-button { margin-right: 0.75em; }
+
+.fa-pencil-alt,
+.fa-trash-alt { 
+  opacity: 0.8;
+  transition: 0.3s;
+}
+
+.fa-pencil-alt:hover,
+.fa-trash-alt:hover { 
+  opacity: 1;
+  font-size: 18px;
+}
+
+.fa-trash-alt { color: rgb(255, 100, 100); }
 
 #to-do-list {
   list-style: none;
@@ -80,7 +93,7 @@ i { pointer-events: none; }
 
 .checkbox-description-container > p { margin: 0.7em; }
 
-.to-do-clear {
+#to-do-clear {
   width: 100%;
   text-align: center;
   border: 0;
@@ -94,3 +107,7 @@ i { pointer-events: none; }
   text-decoration: line-through;
   opacity: 0.5;
 }
+
+.show { display: flex; }
+
+.hidden { display: none; }
